requirements.txt
numpy
torch
torchvision
opencv-python
onnx
onnxruntime-gpu
scikit-image
insightface
requests
kornia
dill
wandb
easydict
albumentations